requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.18.0
fsspec==2025.5.1
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
portalocker==2.10.1
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.1
pytesseract==0.3.13
python-dotenv==1.1.1
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.3
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==4.1.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
